+++ src/pageComponents/Login/LoginIcons.tsx
@@ -1,0 +1,12 @@
+import { ReactComponent as Remove } from "../../assets/icons/remove.svg";
+import { ReactComponent as Hide } from "../../assets/icons/hide.svg";
+import { ReactComponent as Visible } from "../../assets/icons/visible.svg";
+
+import { ReactComponent as Unchecked } from "../../assets/elements/unchecked.svg";
+import { ReactComponent as Checked } from "../../assets/elements/checked.svg";
+
+import { ReactComponent as Google } from "../../assets/icons/google.svg";
+import { ReactComponent as Kakao } from "../../assets/icons/kakaotalk.svg";
+import { ReactComponent as Naver } from "../../assets/icons/naver.svg";
+
+export { Remove, Hide, Visible, Unchecked, Checked, Google, Kakao, Naver };
